backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
pydantic==2.9.2
pydantic-settings==2.5.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.2.0
bcrypt==4.0.1
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.22
openai==1.50.2
httpx==0.27.2

